+++ packages/cli/src/lib/commands/assert/index.ts
@@ -1,7 +1,7 @@
 import  { YargsCommandObject } from '../../core/yargs/types';
 import { logVerbose } from '../../core/loggin/index';
 import { readBudgets } from './utils/budgets';
-import { userFlowReportToMdTable } from './processes/md-table';
+import {generateMdReport} from "../collect/processes/generate-reports";
 import { readFile } from '../../core/file';
 import { getCLIConfigFromArgv } from '../../global/rc-json';
 import { RcJson } from '../../global/rc-json/types';
@@ -14,7 +14,7 @@ export const assertCommand: YargsCommandObject = {
       logVerbose(`run "assert" as a yargs command`);
       const cfg = getCLIConfigFromArgv(argv);
       const json = JSON.parse(readFile('./packages/cli/docs/raw/order-coffee.uf.json').toString());
-      const table = userFlowReportToMdTable(json);
+      const mdReport = generateMdReport(json);
      //  await run(cfg);
     }
   }